+++ src/plugins/codemirror-block/react/components/style.ts
@@ -1,0 +1,7 @@
+import { createStaticStyles } from 'antd-style';
+
+export const styles = createStaticStyles(({ css }) => ({
+  container: css`
+    min-width: 160px;
+  `,
+}));